added stats to planner and reworked the 18 rows into 6 x 3 party members
