+++ src/components/AppHeader.js
@@ -9,7 +9,7 @@ class AppHeader extends PureComponent {
       <header className="app-header">
         <div className="container">
         <div className="app-header-left">
-          <h3>Siralim Planner <span className="version-num">v1.01</span></h3>
+          <h3>Siralim Planner <span className="version-num">v1.02</span></h3>
         </div>
         <div className="app-header-right">
           <p><a href="https://docs.google.com/spreadsheets/d/1qvWwf1fNB5jN8bJ8dFGAVzC7scgDCoBO-hglwjTT4iY/edit#gid=0" target="_blank" rel="noreferrer">Compendium</a> {this.props.compendiumVersion}</p>
