+++ src/markdown/ChangeLog.md
@@ -1,10 +1,15 @@
 ## Changelog
 
-### 25 September 2021
+### v1.02 (1 October 2021)
+
+* Added monster stats to the monster selection table.
+* The monster selection table can now be sorted by clicking on the headings.
+
+### v1.01 (25 September 2021)
 
 * Added creature sprites, sourced from the Siralim Ultimate API.
 * Tool now allows you to upload your build directly from the export from the game.
 
-### 24 September 2021
+### v1.0 (24 September 2021)
 
-Released.
+* Released.
