+++ src/scss/mixins.scss
@@ -1,0 +1,7 @@
+$primary-color: #0063db;
+
+$color-nature: #5a9019;
+$color-chaos: saturate(#b50707, 5);
+$color-sorcery: #9660d1;
+$color-death: #0f1f94;
+$color-life: darken(#ffd373, 25);
